+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -3653,8 +3653,6 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         :return: the id of the station to which this tile is assigned
         """
         station = self._attribute_state["stationId"].read()
-        message = f"stationId: read value = {station}"
-        self.logger.info(message)
         return station
 
     @stationId.write  # type: ignore[no-redef]
@@ -4059,6 +4057,31 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             "To execute this command we must be in state "
             "'Initialised' or 'Synchronised'! "
             f"Tile is currently in state {prog_state}"
+        )
+        tango.Except.throw_exception(reason, msg, self.get_name())
+        return False
+
+    def _is_synchronised(self: MccsTile, *args: Any) -> bool:
+        """
+        Return a flag representing whether we are in Synchronised state.
+
+        :param args: The tango.AttReqType.
+
+        :return: True if Tile is in Synchronised state.
+        """
+        if self.component_manager._initialise_executing:
+            reason = "CommandNotAllowed"
+            msg = "Cannot execute this command while initialise is executing!"
+            tango.Except.throw_exception(reason, msg, self.get_name())
+            return False
+
+        prog_state = self._attribute_state["tileProgrammingState"].read()[0]
+        if prog_state == "Synchronised":
+            return True
+        reason = "CommandNotAllowed"
+        msg = (
+            "To execute this command we must be in state "
+            f"'Synchronised'! Tile is currently in state {prog_state}"
         )
         tango.Except.throw_exception(reason, msg, self.get_name())
         return False
@@ -5861,7 +5884,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             return json.dumps(item_new[0])
         return json.dumps(item_new)
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_initialised",
+    )
     @stb.validators.validate_json_args(schema=SetLmcDownload_SCHEMA)
     def SetLmcDownload(
         self: MccsTile,
@@ -5915,7 +5942,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             self.component_manager.refresh_tile_info()
         return (result_codes, messages)
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_initialised",
+    )
     @stb.validators.validate_json_args(schema=SetLmcIntegratedDownload_SCHEMA)
     def SetLmcIntegratedDownload(
         self: MccsTile,
@@ -6159,7 +6190,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             cls=NumpyEncoder,
         )
 
-    @command(dtype_in="DevVarLongArray", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevVarLongArray",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_initialised",
+    )
     def SetBeamFormerRegions(
         self: MccsTile, argin: list[int]
     ) -> stb.type_hints.DevVarLongStringArrayType:
@@ -6457,7 +6492,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         """
         return self.component_manager.apply_calibration(switch_time)
 
-    @command(dtype_in="DevVarDoubleArray", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevVarDoubleArray",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_synchronised",
+    )
     def LoadPointingDelays(
         self: MccsTile, argin: list[float]
     ) -> stb.type_hints.DevVarLongStringArrayType:
@@ -6504,7 +6543,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
 
         return self.component_manager.load_pointing_delays(delay_array, beam_index)
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_synchronised",
+    )
     def ApplyPointingDelays(
         self: MccsTile, start_time: str
     ) -> stb.type_hints.DevVarLongStringArrayType:
@@ -6750,7 +6793,11 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         """
         return self.component_manager.stop_integrated_data()
 
-    @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
+    @command(
+        dtype_in="DevString",
+        dtype_out="DevVarLongStringArray",
+        fisallowed="_is_synchronised",
+    )
     @stb.validators.validate_json_args(schema=SendDataSamples_SCHEMA)
     def SendDataSamples(
         self: MccsTile,
@@ -7131,7 +7178,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return ([ResultCode.OK], ["StartPatternGenerator command completed OK"])
 
     @engineering_mode_required
-    @command(dtype_out="DevVarLongStringArray")
+    @command(dtype_out="DevVarLongStringArray", fisallowed="_is_synchronised")
     def StartADCs(self: MccsTile) -> stb.type_hints.DevVarLongStringArrayType:
         """
         Start the ADCs.
@@ -7148,7 +7195,7 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         return ([ResultCode.OK], ["StartAdcs command completed OK"])
 
     @engineering_mode_required
-    @command(dtype_out="DevVarLongStringArray")
+    @command(dtype_out="DevVarLongStringArray", fisallowed="_is_synchronised")
     def StopADCs(self: MccsTile) -> stb.type_hints.DevVarLongStringArrayType:
         """
         Stop the ADCs.