+++ tests/unit/station/conftest.py
@@ -594,6 +594,7 @@ def station_self_check_manager_fixture(
         tile_trls=tile_trls,
         subrack_trls=subrack_trls,
         daq_trl="",
+        wren_trl="",
     )
     # Jank to get around https://github.com/python/mypy/issues/3115 and
     # https://github.com/python/mypy/issues/16509
@@ -604,6 +605,7 @@ def station_self_check_manager_fixture(
             tile_trls=list(tile_trls),
             subrack_trls=list(subrack_trls),
             daq_trl="",
+            wren_trl="",
         )
         for tpm_test in [
             PassTest,
@@ -617,6 +619,7 @@ def station_self_check_manager_fixture(
             tile_trls=list(tile_trls),
             subrack_trls=list(subrack_trls),
             daq_trl="",
+            wren_trl="",
         )
         for tpm_test in [
             ErrorTest,